% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metricas.R
\encoding{UTF-8}
\name{metricas}
\alias{metricas}
\title{Métricas de la matriz de confusión}
\usage{
metricas(reales, predichos)
}
\arguments{
\item{reales}{valores reales}

\item{predichos}{valores predichos por el modelo}
}
\value{
\code{Recall}, \code{Precision}, \code{Accuracy} y \verb{F1 Score}
}
\description{
Métricas de la matriz de confusión
}
