% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLM-deprecated.R
\name{MGLM-deprecated}
\alias{MGLM-deprecated}
\alias{ddirm}
\alias{dgdirm}
\alias{dneg}
\alias{rdirm}
\alias{rgdirm}
\title{Deprecated function(s) in the MGLM package}
\usage{
ddirm(...)

rdirm(...)

dgdirm(...)

rgdirm(...)

dneg(Y, alpha, beta)
}
\arguments{
\item{...}{parameters to be passed to the modern version of the function}

\item{Y, alpha, beta}{for functions \code{dnegmn}, note the change in argument order. See Details.}
}
\description{
These functions are provided for compatibility with older version of
the yourPackageName package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{ddirm} \tab now a synonym for \code{\link{ddirmn}}\cr
  \code{dgdirm} \tab now a synonym for \code{\link{dgdirmn}}\cr
  \code{dneg} \tab now a synonym for \code{\link{dnegmn}}\cr
  \code{rdirm} \tab now a synonym for \code{\link{rdirmn}}\cr
  \code{rgdirm} \tab now a synonym for \code{\link{rgdirmn}}\cr
}

The function \code{dneg} has been deprecated. Use \code{dnegmn} instead.

Note the change in argument order: 
\code{dneg(Y, prob, beta)} and \code{dnegmn(Y, alpha, beta)} from MGLM_0.0.8 have been deprecated;
use \code{dnegmn(Y, beta, prob = alpha/(rowSums(alpha)+1), alpha=NULL)} instead.
}

