\name{MGLM-package}
\alias{MGLM-package}
\alias{MGLM}
\docType{package}
\title{
Multivariate response generalized linear models
}
\description{
The package provides functions that (1) fit multivariate discrete distributions, (2) generate random numbers from multivariate discrete distributions, and (3) run regression and penalized regression on the multivariate categorical response data.  Implemented models include: multinomial logit model, Dirichlet multinomial model, generalized Dirichlet multinomial model, and negative multinomial model. Making the best of the minorization-maximization (MM) algorithm and Newton-Raphson method, we derive and implement stable and efficient algorithms to find the maximum likelihood estimates. On a multi-core machine, multi-threading is supported.
}
\details{
\tabular{ll}{
Package: \tab MGLM\cr
Type: \tab Package\cr
Version: \tab 0.0.8\cr
Date: \tab 2017-03-15\cr
License: \tab  GPL (>= 2)\cr
Depends: \tab R (>= 3.0.0), methods, stats, parallel \cr
}
}
\author{
Yiwen Zhang and Hua Zhou
}

