% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaclasses.R
\docType{class}
\name{MGLMreg-class}
\alias{MGLMreg-class}
\title{Class \code{"MGLMreg"}}
\description{
Objects can be created by calls of the form \code{new("MGLMreg", ...)}.
}
\section{Slots}{

\describe{
\item{\code{call}}{object of class \code{"call"}.}

\item{\code{data}}{object of class \code{"list"} ,
consists of both the predictor matrix and the response matrix.}

\item{\code{coefficients}}{object of class \code{"list"} or \code{"matrix"},
the estimated parameters.}

\item{\code{SE}}{object of class \code{"list"} or \code{"matrix"},
the standard errors of the parameters.}

\item{\code{test}}{object of class \code{"matrix"},
the test statistics and p-values.}

\item{\code{Hessian}}{object of class \code{"matrix"},
the Hessian matrix.}

\item{\code{logL}}{object of class \code{"numeric"},
the loglikelihood.}

\item{\code{BIC}}{object of class \code{"numeric"},}

\item{\code{AIC}}{object of class \code{"numeric"},
Akaike information criterion.}

\item{\code{iter}}{object of class \code{"numeric"}, 
the number of iteration used.}

\item{\code{distribution}}{object of class \code{"character"},
the distribution fitted.}

\item{\code{fitted}}{object of class \code{"vector"},
the fitted value.}

\item{\code{gradient}}{object of class \code{"numeric"} or \code{"matrix"},
the gradient at the estimated parameter values.}

\item{\code{wald.value}}{object of class \code{"numeric"} or \code{"logical"},
the Wald statistics.}

\item{\code{wald.p}}{object of class \code{"numeric"} or \code{"logical"},
the p values of Wald test.}

\item{\code{Dof}}{object of class \code{"numeric"}, 
the degrees of freedom.}
}}

\examples{
showClass("MGLMreg")

}
\author{
Yiwen Zhang and Hua Zhou
}
\keyword{classes}
