% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaclasses.R
\docType{class}
\name{MGLMtune-class}
\alias{MGLMtune-class}
\title{Class \code{"MGLMtune"}}
\description{
A class containing the results from the \code{MGLMtune}.
}
\section{Slots}{

\describe{
\item{\code{call}}{object of class \code{"call"}.}

\item{\code{select}}{object of class \code{"MGLMsparsereg"},
regularized regression results given by the optimal tuning parameter.}

\item{\code{path}}{object of class \code{"data.frame"},
the BIC, AIC, log-likelihood and degrees of freedom given each tuning parameter.}

\item{\code{select.list}}{object of class \code{"list"},
the regularized regression results at each tuning grid point.}
}}

\examples{
showClass("MGLMtune")

}
\author{
Yiwen Zhang and Hua Zhou
}
\keyword{classes}
