% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ganancia.R
\encoding{UTF-8}
\name{model.ganancia}
\alias{model.ganancia}
\title{Curva de ganancia}
\usage{
model.ganancia(reales, probabilidades)
}
\arguments{
\item{reales}{valores reales}

\item{probabilidades}{probabilidades de las predicciones}
}
\value{
Gráfico de ganancia
}
\description{
Curva de ganancia
}
