% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS.R
\encoding{UTF-8}
\name{model.ks}
\alias{model.ks}
\title{Métrica de Kolmogorov - Smirnov}
\usage{
model.ks(realestrain, probabilidadestrain, realestest, probabilidadestest)
}
\arguments{
\item{realestrain}{valores reales del conjunto de entrenamiento.}

\item{probabilidadestrain}{probabilidades de los valores predichos del conjunto de entrenamiento.}

\item{realestest}{valores reales del conjunto de prueba.}

\item{probabilidadestest}{probabilidades de los valores predichos del conjunto de prueba.}
}
\value{
Gráfico de discriminación
}
\description{
Métrica de Kolmogorov - Smirnov
}
