% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR.R
\encoding{UTF-8}
\name{model.pr}
\alias{model.pr}
\title{Gráfico de Precision - Recall}
\usage{
model.pr(realestrain, probabilidadestrain, realestest, probabilidadestest)
}
\arguments{
\item{realestrain}{valores reales de entrenamiento.}

\item{probabilidadestrain}{probabilidades de la predicción del entrenamiento.}

\item{realestest}{valores reales de prueba.}

\item{probabilidadestest}{probabilidades de la predicción de la prueba.}
}
\value{
Gráfico de Precision vs Recall.
}
\description{
Gráfico de Precision - Recall
}
