% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\encoding{UTF-8}
\name{model.roc}
\alias{model.roc}
\title{Curva de ROC}
\usage{
model.roc(realestrain, probabilidadestrain, realestest, probabilidadestest)
}
\arguments{
\item{realestrain}{valores reales de entrenamiento.}

\item{probabilidadestrain}{probabilidades de la predicción del entrenamiento.}

\item{realestest}{valores reales de prueba.}

\item{probabilidadestest}{probabilidades de la predicción de la prueba.}
}
\value{
Curva de ROC y sus respectivos AUC.
}
\description{
Curva de ROC
}
