% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{mtrglm}
\alias{mtrglm}
\title{Métricas de GLM}
\usage{
mtrglm(
  y,
  x,
  link = "logit",
  p = 0.8,
  balance = NULL,
  ms = NULL,
  semilla = 9999,
  influencia = TRUE
)
}
\arguments{
\item{y}{variable respuesta}

\item{x}{covariables}

\item{link}{link}

\item{p}{tamaño del entrenamiento}

\item{balance}{tipo de balanceo}

\item{ms}{método de selección de variables}

\item{semilla}{semilla de muestreo}

\item{influencia}{eliminación de los puntos de influecias del modelo}
}
\value{
métricas de desempeño
}
\description{
Métricas de GLM
}
