\name{predict-methods}
\docType{methods}
\alias{predict-methods}
\alias{predict,ANY-method}
\alias{predict,MGLMreg-method}
\title{Predict using the fitted model}
\description{
Predict using the fitted model from \code{MGLMreg} when given a new set of covariates.}
\section{Methods}{
\describe{

\item{\code{signature(object = "ANY")}}{
The new covariates data matrix.
}

\item{\code{signature(object = "MGLMreg")}}{
The model object.
}
}}
\keyword{methods}

