\name{rnaseq}
\docType{data}
\alias{rnaseq}
\title{RNA-seq count data}
\description{
RNA-seq data simulated following the standard procedures
(provided by Dr. Wei Sun, \url{wsun@bios.unc.edu}).
}
\usage{rnaseq}
\format{A data frame containing 10 columns and 100 rows. 
		The first 6 columns are the expression counts of 6 exons of a 
		gene; the last four columns are the covariates:
		age, gender, treatment, and total number of reads.
}
\source{Dr. Sun Wei <\url{wsun@bios.unc.edu}>}
\keyword{dataset}
