% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Puntos.R
\encoding{UTF-8}
\name{scatter.model}
\alias{scatter.model}
\title{Gráfico de puntos para métricas de influencia}
\usage{
scatter.model(data, color, title)
}
\arguments{
\item{data}{vector de valores de la métrica.}

\item{color}{vector con los valores que violan la métrica.}

\item{title}{título del gráfico.}
}
\value{
Se utiliza para los gráficos de características influencia de los residuos.
}
\description{
Gráfico de puntos para métricas de influencia
}
