% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.MGLM.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,MGLMfit-method}
\alias{show,MGLMreg-method}
\alias{show,MGLMsparsereg-method}
\alias{show,MGLMtune-method}
\title{Show an object}
\usage{
\S4method{show}{MGLMfit}(object)

\S4method{show}{MGLMreg}(object)

\S4method{show}{MGLMsparsereg}(object)

\S4method{show}{MGLMtune}(object)
}
\arguments{
\item{object}{an object to be printed. Should be of class \code{"MGLMfit"}, \code{"MGLMreg"},
\code{"MGLMsparsereg"} or \code{"MGLMtune"}.}
}
\description{
Display the object by printing its class.
}
\examples{
library("MGLM")
data("rnaseq")
data <- rnaseq[, 1:6]
gdmFit <- MGLMfit(data, dist = "GDM")
show(gdmFit)
}
