% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaclasses.R
\docType{class}
\name{MGLMfit-class}
\alias{MGLMfit-class}
\title{Class \code{"MGLMfit"}}
\description{
A class containing the model fitting results from the \code{MGLMfit}.
}
\section{Slots}{

\describe{
\item{\code{estimate}}{object of class \code{"vector"}, containing the parameter estimates.}

\item{\code{SE}}{object of class \code{"vector"},
containing the standard errors of the estimates.}

\item{\code{vcov}}{object of class \code{"matrix"},
the variance covariance matrix of the parameter estimates.}

\item{\code{logL}}{object of class \code{"numeric"}, 
the fitted log likelihood.}

\item{\code{BIC}}{object of class \code{"numeric"}, 
Bayesian information criterion.}

\item{\code{AIC}}{object of class \code{"numeric"},
Akaike information criterion.}

\item{\code{LRTpvalue}}{object of class \code{"numeric"},
likelihood ratio test p value.}

\item{\code{gradient}}{object of class \code{"numeric"} or \code{"matrix"},
containing the gradient.}

\item{\code{iter}}{object of class \code{"numeric"}, 
number of iteration used.}

\item{\code{distribution}}{object of class \code{"character"},
the distribution fitted.}

\item{\code{fitted}}{object of class \code{"vector"},
the fitted mean of each category.}

\item{\code{LRT}}{object of class \code{"numeric"}, 
the likelihood ratio test statistic.}
}}

\examples{
showClass("MGLMfit")

}
\author{
Yiwen Zhang and Hua Zhou
}
\keyword{classes}
