% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{mtrglm}
\alias{mtrglm}
\title{Métricas de GLM}
\usage{
mtrglm(
  y,
  x,
  link = "logit",
  p = 0.8,
  balance = NULL,
  ms = NULL,
  semilla = 9999,
  influencia = TRUE
)
}
\arguments{
\item{y}{variable respuesta binaria}

\item{x}{data frame de covariables}

\item{link}{función de enlace ("logit", "probit" o "cloglog"), por defecto es "logit".}

\item{p}{valor entre 0 y 1 que indica la proporción de entrenamiento, por defecto es 0.8.}

\item{balance}{tipo de balanceo ("under" u "over"), por defecto es NULL (sin balancear).}

\item{ms}{método de selección de covariables ("forward" o "backward"), por defecto es NULL (se ocupan todas las covariables).}

\item{semilla}{semilla de muestreo, por defecto es 9999.}

\item{influencia}{eliminación de los puntos de influecias del modelo, por defecto es TRUE.}
}
\value{
métricas de desempeño
}
\description{
Métricas de GLM
}
\examples{
df = iris[, c(1,4,5)]
df$Species = ifelse(df$Species == "virginica", 1, 0)
metrics = mtrglm(df$Species, df[,c(1:2)])
metrics$Train_Metrics
metrics$ROC

}
